+++ br_account/tests/__init__.py
@@ -3,7 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 
-from . import test_taxes
+#from . import test_taxes
 from . import test_account_invoice
-from . import test_fiscal_position
-from . import test_import_ibpt
+#from . import test_fiscal_position
+#from . import test_import_ibpt
